Trying to parallelize sparse state vector
